requirements.txt
# python==3.10
accelerate==1.10.0
anthropic
bitsandbytes==0.44.1
datasets==3.0.1
flash-attn==2.7.2.post1
google-genai
gradio
matplotlib
openai==1.55.3
peft==0.17.0
pillow==10.4.0
qwen_vl_utils==0.0.8
tensorboard
torch==2.4.0
torchvision==0.19.0
trl==0.16.1
tqdm==4.66.5
transformers==4.55.2